rfac: Include loggingMiddleware in server.go
